Prevents data from being saved to Firebase if user toggled tracking off
